自動再生用の基本クラスを追加
