+++ Ragnarok.Forms.Shogi/AutoPlayBase.cs
@@ -9,33 +9,41 @@ namespace Ragnarok.Forms.Shogi
 {
     using View;
 
-    /// <summary>
-    /// 自動再生の種別です。
-    /// </summary>
-    public enum AutoPlayType
+    public interface IAutoPlay
     {
         /// <summary>
-        /// 指し手を動かしません。
+        /// 自動再生を行う将棋エレメントを取得します。
         /// </summary>
-        None,
+        GLShogiElement ShogiElement { get; set; }
+
         /// <summary>
-        /// 与えられた指し手を自動再生します。
+        /// 現在の局面を取得または設定します。
         /// </summary>
-        Normal,
+        Board Board { get; }
+
         /// <summary>
-        /// 局面を元に戻しながら自動再生します。
+        /// 重要な自動再生かどうかを取得または設定します。
         /// </summary>
-        Undo,
+        /// <remarks>
+        /// 真の場合は、GUIのマウス押下で自動再生をキャンセルしません。
+        /// </remarks>
+        bool IsImportant { get; }
+
         /// <summary>
-        /// 局面を次に進めながら自動再生します。
+        /// 更新します。
         /// </summary>
-        Redo,
+        bool Update(TimeSpan elapsed);
+
+        /// <summary>
+        /// 自動再生の途中停止を行います。
+        /// </summary>
+        void Stop();
     }
 
     /// <summary>
     /// 指し手の自動再生時に使われます。再生用の変化を保存します。
     /// </summary>
-    public class AutoPlay
+    public class AutoPlayBase : IAutoPlay
     {
         /// <summary>
         /// 指し手のデフォルトの再生間隔です。
@@ -48,8 +56,6 @@ namespace Ragnarok.Forms.Shogi
         public static readonly TimeSpan DefaultEffectFadeInterval =
             TimeSpan.FromSeconds(0.5);
 
-        private readonly List<Move> moveList;
-        private int moveIndex;
         private IEnumerator<bool> enumerator;
 
         /// <summary>
@@ -72,7 +78,7 @@ namespace Ragnarok.Forms.Shogi
         public GLShogiElement ShogiElement
         {
             get;
-            internal set;
+            set;
         }
 
         /// <summary>
@@ -94,29 +100,12 @@ namespace Ragnarok.Forms.Shogi
         }
 
         /// <summary>
-        /// 局面のコピーに対して指し手を動かすかどうかを取得します。
-        /// </summary>
-        public bool IsCloneBoard
-        {
-            get;
-            private set;
-        }
-
-        /// <summary>
         /// 重要な自動再生かどうかを取得または設定します。
         /// </summary>
         /// <remarks>
         /// 真の場合は、GUIのマウス押下で自動再生をキャンセルしません。
         /// </remarks>
         public bool IsImportant
-        {
-            get { return !IsCloneBoard; }
-        }
-
-        /// <summary>
-        /// 自動再生の種類を取得します。
-        /// </summary>
-        public AutoPlayType AutoPlayType
         {
             get;
             private set;
@@ -183,23 +172,6 @@ namespace Ragnarok.Forms.Shogi
         {
             get;
             set;
-        }
-
-        /// <summary>
-        /// 指し手の最大数を取得します。
-        /// </summary>
-        public int MaxMoveCount
-        {
-            get;
-            set;
-        }
-        
-        /// <summary>
-        /// まだ指し手が残っているか取得します。
-        /// </summary>
-        private bool HasMove
-        {
-            get { return (this.moveIndex < MaxMoveCount); }
         }
 
         /// <summary>
@@ -282,65 +254,11 @@ namespace Ragnarok.Forms.Shogi
         }
 
         /// <summary>
-        /// 指し手を一手だけ進めます。
-        /// </summary>
-        private void DoMove()
-        {
-            if (!HasMove || Board == null)
-            {
-                return;
-            }
-
-            switch (AutoPlayType)
-            {
-                case AutoPlayType.Normal:
-                    var move = this.moveList[this.moveIndex++];
-                    if (move != null)
-                    {
-                        Board.DoMove(move);
-                    }
-                    break;
-                case AutoPlayType.Undo:
-                    this.moveIndex += 1;
-                    Board.Undo();
-                    break;
-                case AutoPlayType.Redo:
-                    this.moveIndex += 1;
-                    Board.Redo();
-                    break;
-            }
-        }
-
-        /// <summary>
         /// 指し手を進めます。
         /// </summary>
-        protected IEnumerable<bool> DoMoveExecutor()
+        protected virtual IEnumerable<bool> DoMoveExecutor()
         {
-            // 最初の指し手はすぐに表示します。
-            DoMove();
-
-            while (HasMove)
-            {
-                if (PositionFromBase > Interval)
-                {
-                    BasePosition += Interval;
-                    DoMove();
-                }
-
-                yield return true;
-            }
-
-            // 必要なら最後の指し手を動かした後に一手分だけ待ちます。
-            // エフェクトを表示するためです。
-            if (IsWaitForLastMove)
-            {
-                while (PositionFromBase < Interval)
-                {
-                    yield return true;
-                }
-
-                BasePosition += Interval;
-            }
+            yield return false;
         }
 
         /// <summary>
@@ -424,9 +342,9 @@ namespace Ragnarok.Forms.Shogi
         /// <summary>
         /// 自動再生の途中停止を行います。
         /// </summary>
-        public void Stop()
+        public virtual void Stop()
         {
-            if (ShogiElement !=null)
+            if (ShogiElement != null)
             {
                 ShogiElement.AutoPlayOpacity = 0.0;
             }
@@ -435,38 +353,14 @@ namespace Ragnarok.Forms.Shogi
             Position = TimeSpan.Zero;
             BasePosition = TimeSpan.Zero;
             this.enumerator = null;
-            this.moveIndex = 0;
 
             RaiseStopped();
         }
 
         /// <summary>
-        /// オブジェクトの妥当性を検証します。
-        /// </summary>
-        public bool Validate()
-        {
-            if (StartBoard == null || !StartBoard.Validate())
-            {
-                return false;
-            }
-
-            if (AutoPlayType == AutoPlayType.Normal)
-            {
-                if (this.moveList == null)
-                {
-                    return false;
-                }
-
-                return StartBoard.CanMoveList(this.moveList);
-            }
-
-            return true;
-        }
-
-        /// <summary>
         /// 共通コンストラクタ
         /// </summary>
-        private AutoPlay(Board board, bool isCloneBoard)
+        public AutoPlayBase(Board board, bool isImportant)
         {
             if (board == null)
             {
@@ -475,8 +369,8 @@ namespace Ragnarok.Forms.Shogi
 
             UpdateEnumeratorFactory = () => GetUpdateEnumerator();
             StartBoard = board;
-            Board = (isCloneBoard ? board.Clone() : board);
-            IsCloneBoard = isCloneBoard;
+            Board = board;
+            IsImportant = isImportant;
             Interval = DefaultInterval;
             EffectFadeInterval = DefaultEffectFadeInterval;
             BeginningInterval = TimeSpan.Zero;
@@ -484,48 +378,6 @@ namespace Ragnarok.Forms.Shogi
             Position = TimeSpan.Zero;
             BasePosition = TimeSpan.Zero;
             IsWaitForLastMove = true;
-        }
-
-        /// <summary>
-        /// コンストラクタ
-        /// </summary>
-        public AutoPlay(Board board, bool isCloneBoard,
-                        IEnumerable<Move> moveList)
-            : this(board, isCloneBoard)
-        {
-            if (moveList == null)
-            {
-                throw new ArgumentNullException("moveList");
-            }
-
-            AutoPlayType = AutoPlayType.Normal;
-            MaxMoveCount = moveList.Count();
-
-            this.moveList = new List<Move>(moveList);
-        }
-
-        /// <summary>
-        /// コンストラクタ
-        /// </summary>
-        public AutoPlay(Board board, bool isCloneBoard,
-                        AutoPlayType autoPlayType, int maxMoveCount = -1)
-            : this(board, isCloneBoard)
-        {
-            if (autoPlayType != AutoPlayType.Undo &&
-                autoPlayType != AutoPlayType.Redo)
-            {
-                throw new ArgumentException(
-                    "アンドゥかリドゥを選択してください。",
-                    "autoPlayType");
-            }
-
-            AutoPlayType = autoPlayType;
-            MaxMoveCount = (
-                maxMoveCount >= 0 ?
-                maxMoveCount :
-                (autoPlayType == AutoPlayType.Undo ?
-                    board.CanUndoCount :
-                    board.CanRedoCount));
         }
     }
 }
